requirements.txt
asgiref==3.8.1
channels==4.0.0
click==8.1.7
Django==4.2.10
djangorestframework==3.14.0
dnspython==2.6.1
et-xmlfile==1.1.0
h11==0.14.0
openpyxl==3.1.3
pymongo==4.7.2
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
uvicorn==0.27.1
